requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Data visualization and monitoring
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: For advanced GPU support
# torch-tensorrt  # Uncomment for TensorRT optimization
